LayoutTests/fast/events/message-port-expected.txt

Test cross-frame MessagePort communication.

Should be a series of SUCCESS messages, followed with DONE.

Simple exchange: SUCCESS
Posted port: SUCCESS
Conversation: SUCCESS
A port is active: SUCCESS
A closed port is inactive: SUCCESS
DONE
